backend/pyproject.toml
[project]
name = "api-analytics-service"
version = "1.0.0"
description = "Backend for API Analytics"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.1",
    "fastapi[standard]>=0.128.0",
    "sqlalchemy>=2.0.45",
    "slowapi>=0.1.9",
    "redis>=5.0.0",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.10.1",
]
